Corrected the OrbitBehavior, added simple test to the Game class.
Added a trivial steering behavior

Temporarily, Physics updating is taken care of in the GameObject.

This is terrible, but I didn't want to write the entire physics system updating and world structure just to test orbit behaviors.
MAKE SURE THAT IS CORRECTED WHEN THINGS LOOK WEIRD.
I suspect that this will cause debug issues when collision resolution starts, so once I move into that realm, I need to correct the model.

Minor success note, Rendering 800 orbs seems to be handled smoothly, even given the primitive rendering technique for borders.  Go me!

+++ src/Math/Vec2.java
@@ -73,7 +73,7 @@ public class Vec2 {
 
 // -------------------------- OTHER METHODS --------------------------
 
-    public void Add(float c) {
+    public void Add(double c) {
         X += c;
         Y += c;
     }
